refactor(settings): extract SidebarSection to sections/
Moves the Sidebar tab — visible sections (folders, backlinks), note
sort order, and left/right panel widths — into its own file.

No behavior changes.
